SFEperl-test-differences: build package for perl-522
--HG--
branch : spec_SFEperl-test-differences

+++ specs/SFEperl-test-differences.spec
@@ -4,7 +4,8 @@
 %define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define enable_test %( if [ "x${PERL_DISABLE_TEST}" = 'xtrue' ]; then echo '0'; else echo '1'; fi )
 %define include_executable 0
 
 %define cpan_name Test-Differences
@@ -30,6 +31,12 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-584
 Summary:          Test strings and data structures and show differences if not ok
 BuildRequires:    runtime/perl-584 = *
 BuildRequires:    library/perl-5/extutils-makemaker-584
+%if %{enable_test}
+BuildRequires:    library/perl-5/capture-tiny-584
+BuildRequires:    library/perl-5/data-dumper-584
+BuildRequires:    library/perl-5/test-simple-584
+BuildRequires:    library/perl-5/text-diff-584
+%endif
 Requires:         runtime/perl-584 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/capture-tiny-584
@@ -47,6 +54,10 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-510
 Summary:          Test strings and data structures and show differences if not ok
 BuildRequires:    runtime/perl-510 = *
 BuildRequires:    library/perl-5/extutils-makemaker-510
+BuildRequires:    library/perl-5/capture-tiny-510
+BuildRequires:    library/perl-5/data-dumper-510
+BuildRequires:    library/perl-5/test-simple-510
+BuildRequires:    library/perl-5/text-diff-510
 Requires:         runtime/perl-510 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/capture-tiny-510
@@ -64,6 +75,12 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-512
 Summary:          Test strings and data structures and show differences if not ok
 BuildRequires:    runtime/perl-512 = *
 BuildRequires:    library/perl-5/extutils-makemaker-512
+%if %{enable_test}
+BuildRequires:    library/perl-5/capture-tiny-512
+BuildRequires:    library/perl-5/data-dumper-512
+BuildRequires:    library/perl-5/test-simple-512
+BuildRequires:    library/perl-5/text-diff-512
+%endif
 Requires:         runtime/perl-512 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/capture-tiny-512
@@ -81,6 +98,13 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-516
 Summary:          Test strings and data structures and show differences if not ok
 BuildRequires:    runtime/perl-516 = *
 BuildRequires:    library/perl-5/extutils-makemaker-516
+Requires:         library/perl-5/%{ips_cpan_name}
+%if %{enable_test}
+BuildRequires:    library/perl-5/capture-tiny-516
+BuildRequires:    library/perl-5/data-dumper-516
+BuildRequires:    library/perl-5/test-simple-516
+BuildRequires:    library/perl-5/text-diff-516
+%endif
 Requires:         runtime/perl-516 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/capture-tiny-516
@@ -92,27 +116,33 @@ Requires:         library/perl-5/text-diff-516
 Test strings and data structures and show differences if not ok
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{ips_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          Test strings and data structures and show differences if not ok
-BuildRequires:    runtime/perl-520 = *
-BuildRequires:    library/perl-5/extutils-makemaker-520
-Requires:         runtime/perl-520 = *
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/extutils-makemaker-522
+%if %{enable_test}
+BuildRequires:    library/perl-5/capture-tiny-522
+BuildRequires:    library/perl-5/data-dumper-522
+BuildRequires:    library/perl-5/test-simple-522
+BuildRequires:    library/perl-5/text-diff-522
+%endif
+Requires:         runtime/perl-522 = *
 Requires:         library/perl-5/%{ips_cpan_name}
-Requires:         library/perl-5/capture-tiny-520
-Requires:         library/perl-5/data-dumper-520
-Requires:         library/perl-5/test-simple-520
-Requires:         library/perl-5/text-diff-520
+Requires:         library/perl-5/capture-tiny-522
+Requires:         library/perl-5/data-dumper-522
+Requires:         library/perl-5/test-simple-522
+Requires:         library/perl-5/text-diff-522
 
-%description 520
+%description 522
 Test strings and data structures and show differences if not ok
 %endif
 
 
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
@@ -125,8 +155,12 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    [ x${test} = 'xwithout_test' ] || make test
+
+    export CC='cc -m32'
+    export LD='cc -m32'
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null || (export CC='cc -m64'; export LD='cc -m64')
+    make CC="${CC}" LD="${LD}"
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC="${CC}" "LD=${LD}"
     make pure_install
 }
 
@@ -141,7 +175,7 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ x${test} = 'xwithout_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build clean
 }
@@ -158,7 +192,7 @@ modify_bin_dir() {
     then
         for i in $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/*
         do
-            sed -i.bak -e "s/\/usr\/bin\/env ruby/\/usr\/perl5\/${perl-ver}\/bin\/ruby/" ${i}
+            sed -i.bak -e "s!/usr/bin/env perl!/usr/perl5/${perl-ver}/bin/perl!" ${i}
             [ -f ${i}.bak] || rm -f ${i}.bak
         done
     fi
@@ -215,8 +249,8 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
@@ -277,18 +311,19 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Thu Jun 08 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- build package for perl-522
 * Thu Mar 31 2016 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - bump to 0.64
 * Fri Jun 29 2012 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
